requirements.txt
asgiref==3.7.2
coverage==7.4.2
Django==4.2.5
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
